+++ src/pages/Match/Search.tsx
@@ -36,7 +36,7 @@ const SearchMatch: FC = () => {
   } = useInfiniteQuery(['postData', sort, subject], ({ pageParam = 0 }) => fetchPostList(pageParam), {
     getNextPageParam: (lastPage) => (!lastPage.last ? lastPage.nextPage : undefined),
   });
-  console.log(postList);
+
   useEffect(() => {
     if (inView) fetchNextPage();
   }, [inView]);
