+++ src/pages/New/New.tsx
@@ -1,16 +1,10 @@
-import React, { FC, useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../../redux/store';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { apis } from '../../apis';
-import { PostEditDataProps, ImageType } from '../../typings';
+import React, { FC } from 'react';
+import { ImageType } from '../../typings';
+import { usePost } from '../../hooks/post/usePost';
 import { IoMdCloseCircleOutline } from 'react-icons/io';
-import { useQueryClient } from '@tanstack/react-query';
 import MapContainer from '../../components/Modal/MapContainer';
 import Calendars from '../../components/Modal/Calendar';
 import CustomSubject from '../../components/Select/CustomSelect';
-import { usePost } from '../../hooks/post/usePost';
 
 const Newpost: FC = () => {
   const {
@@ -24,114 +18,20 @@ const Newpost: FC = () => {
     onSaveFiles,
     images,
     clearAll,
-    uploadImage,
     imgUrl,
     data,
+    register,
+    handleDataUpload,
+    handleEditUpload,
+    address,
+    subject,
+    modalShow,
+    calendarShow,
+    subjectShow,
+    date,
   } = usePost('');
-  // const location = useLocation();
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-  const { register, getValues } = useForm({});
-  // const [imgUrl, setImageUrl] = useState([]);
-  const address = useSelector((state: RootState) => state.address);
-  const subject = useSelector((state: RootState) => state.subject);
-  const modalShow = useSelector((state: RootState) => state.toggle.modalShow);
-  const calendarShow = useSelector((state: RootState) => state.toggle.calendarShow);
-  const subjectShow = useSelector((state: RootState) => state.toggle.subjectShow);
-  const date = useSelector((state: RootState) => state.calendar.date);
-  // const data = location.state as PostEditDataProps;
+
   clearAll();
-
-  const handleDataUpload = async () => {
-    if (getValues().title.length < 1) {
-      return alert('제목을 입력해주세요.');
-    } else if (subject.subject === '종목을 선택해주세요.') {
-      return alert('종목을 선택해주세요.');
-    } else if (date === '모집 마감일을 선택해 주세요.') {
-      return alert('모집 마감일을 선택해 주세요.');
-    } else if (address.address === '주소를 선택해 주세요.' && address.lat === 0 && address.lng === 0) {
-      return alert('주소를 선택해주세요.');
-    } else if (address.address === undefined) {
-      return alert('올바른 주소를 선택해주세요.');
-    } else if (getValues().content.length < 1) {
-      return alert('내용을 입력해주세요.');
-    } else {
-      const postData = {
-        title: getValues().title,
-        matchDeadline: date,
-        subject: subject.value,
-        content: getValues().content,
-        lat: address.lat,
-        lng: address.lng,
-        address: address.address,
-      };
-      const value = await apis.postUpload(postData);
-      if (uploadImage.length > 0) {
-        const formData = new FormData();
-        for (let i = 0; i < uploadImage.length; i++) {
-          formData.append('files', uploadImage[i]);
-        }
-        await apis.uploadImage(value, formData);
-      }
-      queryClient.removeQueries(['postData']);
-      navigate('/search');
-    }
-  };
-
-  const handleEditUpload = async () => {
-    if (getValues().title.length < 1) {
-      return alert('제목을 입력해주세요.');
-    } else if (subject.subject === '종목을 선택해주세요.') {
-      return alert('종목을 선택해주세요.');
-    } else if (date === '모집 마감일을 선택해 주세요.') {
-      return alert('모집 마감일을 선택해 주세요.');
-    } else if (address.address === '주소를 선택해 주세요.' && address.lat === 0 && address.lng === 0) {
-      return alert('주소를 선택해주세요.');
-    } else if (address.address === undefined) {
-      return alert('올바른 주소를 선택해주세요.');
-    } else if (getValues().content.length < 1) {
-      return alert('내용을 입력해주세요.');
-    } else {
-      const postData = {
-        title: getValues().title,
-        matchDeadline: date,
-        subject: subject.value,
-        content: getValues().content,
-        lat: address.lat,
-        lng: address.lng,
-        address: address.address,
-      };
-      await apis.updatePost(data.postId, postData);
-      if (uploadImage.length > 0) {
-        const formData = new FormData();
-        for (let i = 0; i < uploadImage.length; i++) {
-          if (uploadImage[i] !== null) {
-            formData.append('files', uploadImage[i]);
-          }
-        }
-        await apis.uploadImage(data.postId, formData);
-      }
-      queryClient.removeQueries(['postData']);
-      navigate('/search');
-    }
-  };
-
-  // const handledeletePrevImg = async (id: number) => {
-  //   setImages(images.filter((_, index) => index !== id));
-  //   setUploadImage([...uploadImage.slice(0, id), ...uploadImage.slice(id + 1)]);
-  // };
-
-  // const handledeleteImage = async (id: number) => {
-  //   if (window.confirm('이미지를 삭제하시겠습니까?')) {
-  //     const imgpaths = data.imgpaths[id];
-  //     if (imgpaths !== undefined) {
-  //       await apis.deleteImage(imgpaths['path']);
-  //     }
-  //     if (data.imgurls.length > 0) {
-  //       setImageUrl(imgUrl.filter((_, index) => index !== id));
-  //     }
-  //   }
-  // };
 
   return (
     <section className='flex flex-col w-[100%] h-full'>
